Setup application page for launch
