+++ pages/blog/[slug].js
@@ -94,6 +94,6 @@ export async function getStaticPaths() {
 
   return {
     paths,
-    fallback: false
+    fallback: 'blocking'
   };
 }
